fix: adding releaseId
