+++ packages/core/src/experimental/searchServerSideFunctions/getStaticProps.ts
@@ -47,6 +47,7 @@ export const getStaticProps: GetStaticProps<
           previewData,
           documentId: page.documentId,
           versionId: page.versionId,
+          releaseId: page.releaseId,
         }),
         globalSectionsPromise,
         globalSectionsHeaderPromise,
